benchmark: move mcp-server/lib contents to serving/lib
