+++ serving/benchmarks/rag/benchmark-latency.ts
@@ -1,5 +1,5 @@
-import { Embedder } from "../../mcp-server/lib/embedder.ts";
-import { TfjsEmbedder } from "../../mcp-server/lib/tfjs-embedder.ts";
+import { Embedder } from "../../lib/embedder.ts";
+import { TfjsEmbedder } from "../../lib/tfjs-embedder.ts";
 import fs from "fs";
 import path from "path";
